don't produce spurious symbols
this closes #5

the problem was with creating a new "new symbol" each time mouse button went
up, with invisible old "new symbol" being inserted in the score scene and
saved in the time. This resulted in a new invisible symbol each time user
dragged an existing symbol. Pluse the not yet realized "new symbol" was
saved as well.