find plugins on windows
even with plugins installed, their directory has to be added to the library
search path This time I actually could reproduce the original problem by
temporarily renaming the Qt SDK directory so the application won't find the
plugins there. This patch adds a directory where we install plugins to the
search path at start up, before the main window is created.

Hopefully this should fix the #16.