notice sample changes
We just look at sample timestamps. Simply reloading is not enough - we
have to make sure the model (project) lets views know that it changed
and views refresh sample representations.

Two views will have to be updated: sequencer timeline and waveform (if
the latter displays a sample that has changed)

Re: #38